requirements.txt
# ===============================
# BOT DISCORD GAME NÔNG TRẠI
# Requirements for Ubuntu Server Deployment
# ===============================

# === DISCORD BOT CORE ===
discord.py>=2.3.0              # Discord bot framework chính
aiosqlite>=0.17.0               # Async SQLite database driver
python-dotenv>=0.19.0           # Environment variables management

# === HTTP CLIENTS & API ===
requests>=2.28.0                # Synchronous HTTP requests (Weather API)
aiohttp>=3.8.0                  # Async HTTP client (Gemini API calls)
asyncio-throttle>=1.0.2         # Rate limiting for API calls

# === FILE & DATA HANDLING ===
aiofiles>=0.8.0                 # Async file operations
pathlib                         # File path handling (built-in Python 3.4+)

# === AI & MACHINE LEARNING ===
google-genai>=1.0.0             # Google Gemini AI SDK

# === DATE & TIME ===
# datetime, timedelta - Built-in Python modules
pytz>=2023.3                    # Timezone handling for global users

# === LOGGING & MONITORING ===
# logging - Built-in Python module
colorlog>=6.7.0                 # Colored logging for better readability

# === ASYNC & CONCURRENCY ===
# asyncio - Built-in Python module

# === JSON & DATA PROCESSING ===
# json - Built-in Python module

# === SYSTEM & OS ===
# os, sys, signal - Built-in Python modules
psutil>=5.9.0                   # System monitoring (memory, CPU usage)

# === DEVELOPMENT & TESTING (Optional) ===
# pytest>=7.0.0                # For running tests
# pytest-asyncio>=0.20.0       # Async testing support

# ===============================
# UBUNTU SERVER SPECIFIC
# ===============================

# === PRODUCTION DEPLOYMENT ===
supervisor>=4.2.0               # Process management for bot auto-restart
systemd-python>=234             # Systemd integration (Ubuntu service)

# === DATABASE BACKUP ===
croniter>=1.3.0                 # Cron job scheduling for backups

# === SECURITY ===
cryptography>=40.0.0            # Secure token handling

# ===============================
# VERSION COMPATIBILITY
# ===============================
# Python 3.8+ required
# Ubuntu 20.04+ recommended
# Discord.py 2.3+ for latest Discord features
# Google Gemini API requires internet connection 